ONE-191: Response to code review
