--- src/ONE.Models.CSharp/Common/Configuration/ConfigurationTypeConstants.cs
@@ -1,35 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace ONE.Models.CSharp.Common.Configuration
-{
-    public static class ConfigurationTypeConstants
-    {
-        public static class Dashboards
-        {
-            public const string Id = "9a4a09ca-cf6c-40c0-bc21-1b8f5d306e0a";
-            public static readonly Guid Guid = new Guid(Id);
-        }
-        public static class Graphs
-        {
-            public const string Id = "24ecb22b-aee3-4bac-a88d-9e10402363ba";
-            public static readonly Guid Guid = new Guid(Id);
-        }
-        public static class Worksheets
-        {
-            public const string Id = "bedc5ff2-bc8e-4916-9560-ccc28701d792";
-            public static readonly Guid Guid = new Guid(Id);
-        }
-        public static class IngestClient
-        {
-            public const string Id = "e144a823-2671-4004-a977-4e0092ffd1dc";
-            public static readonly Guid Guid = new Guid(Id);
-        }
-        public static class DataSourceDriver
-        {
-            public const string Id = "e47b66c7-1942-4f9c-b817-e9225c478ea2";
-            public static readonly Guid Guid = new Guid(Id);
-        }
-    }
-}
