--- src/ONE.Models.CSharp/Common/Configuration/Graphs/IGraphAxisMode.cs
@@ -1,9 +1,0 @@
-﻿namespace ONE.Models.CSharp.Models.Common.Configuration.Graphs
-{
-    public enum IGraphAxisMode
-    {
-        PerSeries = 0,
-        AllOnPrimary = 1,
-        LeftOrRight = 2
-    }
-}
