--- src/ONE.Models.CSharp/Common/Configuration/Spreadsheet/ColumnConfig.cs
@@ -1,10 +1,0 @@
-﻿using System;
-
-namespace ONE.Models.CSharp.Models.Common.Configuration.Spreadsheet
-{
-    [Serializable]
-    public class ColumnConfig
-    {
-        public uint columnWidth { get; set; }
-    }
-}
