--- src/ONE.Models.CSharp/Common/Configuration/Spreadsheet/ColumnHeader.cs
@@ -1,16 +1,0 @@
-﻿using ONE.Models.CSharp.Enums;
-
-namespace ONE.Models.CSharp.Models.Common.Configuration.Spreadsheet
-{
-    public class ColumnHeader
-    {
-        public uint id { get; set; }
-        public EnumHeaderType HeaderType
-        {
-            get
-            {
-                return EnumHeaderType.column;
-            }
-        }
-    }
-}
