+++ src/ONE.Models.CSharp/Constants/RoleConstants.cs
@@ -1,0 +1,33 @@
+﻿using System;
+
+namespace ONE.Models.CSharp.Constants
+{
+    public static class RoleConstants
+    {
+        public static class SupportAdmin
+        {
+            public const string Id = "acf28828-4b3b-4f97-9a81-0f02d9a5c84d";
+            public static readonly Guid Guid = new Guid(Id);
+        }
+        public static class Basic
+        {
+            public const string Id = "bbcd2319-3c56-4511-bbef-f2d2d8707748";
+            public static readonly Guid Guid = new Guid(Id);
+        }
+        public static class Admin
+        {
+            public const string Id = "d98c0b6f-2a97-4ed1-a235-7fb55623c2d0";
+            public static readonly Guid Guid = new Guid(Id);
+        }
+        public static class SupportBasic
+        {
+            public const string Id = "72770955-7cd1-4ef8-8e9b-27e8832f1b1a";
+            public static readonly Guid Guid = new Guid(Id);
+        }
+        public static class SuperAdmin
+        {
+            public const string Id = "e927e85a-a0b7-4a15-9c12-84b747588f67";
+            public static readonly Guid Guid = new Guid(Id);
+        }
+    }
+}
