--- src/ONE.Models.CSharp/Enterprise/Twin/Wizard/Location.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace ONE.Models.CSharp.Enterprise.Twin.Wizard
-{
-    public class Location
-    {
-        // Is this a default in the wizard
-        public bool d { get; set; }
-        // Is it possible to have multiple of these locations?
-        public bool m { get; set; }
-        // The id of the digital twin subtype
-        public string id { get; set; }
-    }
-}
