--- src/ONE.Models.CSharp/Enums/EnumHeaderType.cs
@@ -1,11 +1,0 @@
-﻿using System;
-
-namespace ONE.Models.CSharp.Enums
-{
-    [Serializable]
-    public enum EnumHeaderType
-    {
-        column = 1,
-        group = 2
-    }
-}
