--- src/ONE.Models.CSharp/Enums/EnumRelativeDateRange.cs
@@ -1,15 +1,0 @@
-﻿namespace ONE.Models.CSharp.Enums
-{
-    public enum EnumRelativeDateRange
-    {
-        Last7Days = 0,
-        MonthToDate = 1,
-        LastMonth = 2,
-        YearToDate = 3,
-        LastYear = 4,
-        LastQuarter = 5,
-        LastSemiAnnual = 6,
-        Last12Months = 7
-
-    }
-}
