--- src/ONE.Models.CSharp/Enums/EnumSeriesAxis.cs
@@ -1,8 +1,0 @@
-﻿namespace ONE.Models.CSharp.Enums
-{
-    public enum EnumSeriesAxis
-    {
-        Left = 0,
-        Right = 1
-    }
-}
