--- src/ONE.Models.CSharp/Enums/EnumSeriesType.cs
@@ -1,14 +1,0 @@
-﻿namespace ONE.Models.CSharp.Enums
-{
-    public enum EnumSeriesType
-    {
-        area,
-        bar,
-        line,
-        bullet,
-        scatter,
-        bubble,
-        stock,
-        range
-    }
-}
