+++ src/ONE.Models.CSharp/Imposed/Enums/EnumAxisMode.cs
@@ -1,0 +1,9 @@
+﻿namespace ONE.Models.CSharp.Imposed.Enums
+{
+    public enum EnumAxisMode
+    {
+        PerSeries = 0,
+        AllOnPrimary = 1,
+        LeftOrRight = 2
+    }
+}
