+++ src/ONE.Models.CSharp/Imposed/Enums/EnumHeaderType.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace ONE.Models.CSharp.Imposed.Enums
+{
+    [Serializable]
+    public enum EnumHeaderType
+    {
+        column = 1,
+        group = 2
+    }
+}
