+++ src/ONE.Models.CSharp/Imposed/Enums/EnumRelativeDateRange.cs
@@ -1,0 +1,15 @@
+﻿namespace ONE.Models.CSharp.Imposed.Enums
+{
+    public enum EnumRelativeDateRange
+    {
+        Last7Days = 0,
+        MonthToDate = 1,
+        LastMonth = 2,
+        YearToDate = 3,
+        LastYear = 4,
+        LastQuarter = 5,
+        LastSemiAnnual = 6,
+        Last12Months = 7
+
+    }
+}
