+++ src/ONE.Models.CSharp/Imposed/Enums/EnumSeriesAxis.cs
@@ -1,0 +1,8 @@
+﻿namespace ONE.Models.CSharp.Imposed.Enums
+{
+    public enum EnumSeriesAxis
+    {
+        Left = 0,
+        Right = 1
+    }
+}
