+++ src/ONE.Models.CSharp/Imposed/Enums/EnumSeriesType.cs
@@ -1,0 +1,14 @@
+﻿namespace ONE.Models.CSharp.Imposed.Enums
+{
+    public enum EnumSeriesType
+    {
+        area,
+        bar,
+        line,
+        bullet,
+        scatter,
+        bubble,
+        stock,
+        range
+    }
+}
