+++ src/ONE.Models.CSharp/Imposed/Enums/EnumUiElement.cs
@@ -1,0 +1,15 @@
+﻿namespace ONE.Models.CSharp.Imposed.Enums
+{
+    public enum EnumUiElement
+    {
+        Unknown = 0,
+        Textbox = 1,
+        DatePicker = 2,
+        DropDown = 3,
+        List = 4,
+        Options = 5,
+        Form = 6,
+        MeasurementDefinition = 7,
+        Checkbox = 8
+    }
+}
