+++ src/ONE.Models.CSharp/Imposed/Graphs/GraphDataSeries.cs
@@ -1,0 +1,12 @@
+﻿namespace ONE.Models.CSharp.Imposed.Graphs
+{
+    public class GraphDataSeries
+    {
+        public int twinId { get; set; }
+        public string color { get; set; }
+        public bool showLimits { get; set; }
+        public string type { get; set; }
+        public int axis { get; set; }
+
+    }
+}
