+++ src/ONE.Models.CSharp/Imposed/Internationalization/I18NKey.cs
@@ -1,0 +1,22 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace ONE.Models.CSharp.Imposed.Internationalization
+{
+    [Serializable]
+    public class I18NKey
+    {
+        public I18NKey(string module, string type, string key, string value)
+        {
+            Module = module;
+            Type = type;
+            Key = key;
+            Value = value;
+        }
+        public string Module { get; set; }
+        public string Type { get; set; }
+        public string Key { get; set; }
+        public string Value { get; set; }
+    }
+}
