+++ src/ONE.Models.CSharp/Imposed/TwinUiDefinition/I18NKeyTextReference.cs
@@ -1,0 +1,7 @@
+﻿namespace ONE.Models.CSharp.Imposed.TwinUiDefinition
+{
+    public class I18NKeyTextReference
+    {
+        public string i18nKey { get; set; }
+    }
+}
