+++ src/ONE.Models.CSharp/Imposed/TwinUiDefinition/UiAttributeOption.cs
@@ -1,0 +1,8 @@
+﻿namespace ONE.Models.CSharp.Imposed.TwinUiDefinition
+{
+    public class UiAttributeOption
+    {
+        public string value { get; set; }
+        public int key { get; set; }
+    }
+}
