+++ src/ONE.Models.CSharp/Imposed/TwinUiDefinition/UiDefinitionValidator.cs
@@ -1,0 +1,10 @@
+﻿namespace ONE.Models.CSharp.Imposed.TwinUiDefinition
+{
+    public class UiDefinitionValidator
+    {
+        public bool required { get; set; }
+        public int minLength { get; set; }
+        public int maxLength { get; set; }
+
+    }
+}
