+++ src/ONE.Models.CSharp/Imposed/Wizard/TwinPropertyBag.cs
@@ -1,0 +1,14 @@
+﻿using System.Collections.Generic;
+
+namespace ONE.Models.CSharp.Imposed.Wizard
+{
+    public class TwinPropertyBag
+    {
+        //Twin Types to be used
+        public string ty { get; set; }
+        // List of Suggested Locations for this Digital Twin Sub Type
+        public List<WizardLocation> ls { get; set; }
+        // List of Suggested Parameters
+        public List<Parameter> ps { get; set; }
+    }
+}
