+++ src/ONE.Models.CSharp/Imposed/WorksheetView/ColumnConfig.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace ONE.Models.CSharp.Imposed.WorksheetView
+{
+    [Serializable]
+    public class ColumnConfig
+    {
+        public uint columnWidth { get; set; }
+    }
+}
