--- src/ONE.Models.CSharp/UI/I18NKeyTextReference.cs
@@ -1,7 +1,0 @@
-﻿namespace ONE.Models.CSharp.UI
-{
-    public class I18NKeyTextReference
-    {
-        public string i18nKey { get; set; }
-    }
-}
