--- src/ONE.Models.CSharp/UI/Option.cs
@@ -1,8 +1,0 @@
-﻿namespace ONE.Models.CSharp.UI
-{
-    public class Option
-    {
-        public string value { get; set; }
-        public int key { get; set; }
-    }
-}
