--- src/ONE.Models.CSharp/UI/Validator.cs
@@ -1,10 +1,0 @@
-﻿namespace ONE.Models.CSharp.UI
-{
-    public class Validator
-    {
-        public bool required { get; set; }
-        public int minLength { get; set; }
-        public int maxLength { get; set; }
-
-    }
-}
